Test skeletons to prepare for refactoring.
Added some tests to getting and setting user preferences. The exceptional cases are not covered until I find a convinient way to test them.

Additonal changes:

- Reorganized test package structure.
- DbSpec is extends BeforeAfter trait, so every test case runs with a clean db.
